A satisfactory amount of refactoring
for now

+++ src/MatchScore.kt
@@ -1,4 +1,4 @@
 data class MatchScore (
         val csvIndex: Int,
-        val score: Double
+        val score: Float
 )
